+++ src/logic/install.ts
@@ -363,29 +363,39 @@ export type ProfileResponse = {
   ast: ProfileDocumentNode;
 };
 export async function getProfileFromStore(
-  profileId: string,
+  profileId: ProfileId,
   version?: string,
   options?: { logCb?: LogCallback; warnCb?: LogCallback }
 ): Promise<ProfileResponse | undefined> {
-  if (version !== undefined) {
-    profileId = `${profileId}@${version}`;
-  }
+  const profileIdStr = profileId.withVersion(version);
 
   // fetch the profile
   let info: ProfileInfo;
   let profile: string;
   let ast: ProfileDocumentNode;
-  options?.logCb?.(`\nFetching profile ${profileId} from the Store`);
+  options?.logCb?.(`\nFetching profile ${profileIdStr} from the Store`);
 
   try {
-    info = await fetchProfileInfo(profileId);
-    options?.logCb?.(`GET Profile Info ${profileId}`);
+    info = await fetchProfileInfo(profileIdStr);
+    options?.logCb?.(`Fetching profile info of profile ${profileIdStr}`);
 
-    profile = await fetchProfile(profileId);
-    options?.logCb?.(`GET Profile Source File ${profileId}`);
+    profile = await fetchProfile(profileIdStr);
+    options?.logCb?.(`Fetching profile source file for ${profileIdStr}`);
 
-    ast = await fetchProfileAST(profileId);
-    options?.logCb?.(`GET compiled Profile ${profileId}`);
+    try {
+      //This can fail due to validation issues, ast and parser version issues
+      ast = await fetchProfileAST(profileIdStr);
+      options?.logCb?.(`Fetching compiled profile for ${profileIdStr}`);
+    } catch (error) {
+      options?.warnCb?.(
+        `Fetching compiled profile for ${profileIdStr} failed, trying to parse source file`
+      );
+      //We try to parse profile on our own
+      ast = await Parser.parseProfile(profile, profileIdStr, {
+        profileName: profileId.name,
+        scope: profileId.scope,
+      });
+    }
   } catch (error) {
     // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
     options?.warnCb?.(`Could not fetch ${profileId}: ${error}`);
@@ -406,7 +416,7 @@ async function fetchStoreRequestCheckedOrDeferred(
   options?: InstallOptions
 ): Promise<StoreRequestFetched | undefined> {
   const fetched = await getProfileFromStore(
-    request.profileId.id,
+    request.profileId,
     request.version,
     options
   );