+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - prepare integration parameters during configure command
+- install command fallbacks to parse profile source on profile AST validation fail
 
 ## [0.0.26] - 2021-10-13
 ### Added